Image import: Move importer package to common/disk
