+++ cli_tools/common/disk/importer/bootable_disk_processor.go
