+++ cli_tools/common/disk/importer/inflater_test.go
@@ -276,7 +276,7 @@ func TestCreateDaisyInflater_File_NotUEFI(t *testing.T) {
 
 func createDaisyInflaterSafe(t *testing.T, args ImportArguments,
 	inspector imagefile.Inspector) *daisyInflater {
-	args.WorkflowDir = "../../../daisy_workflows"
+	args.WorkflowDir = "../../../../daisy_workflows"
 	inflater, err := NewDaisyInflater(args, inspector)
 	assert.NoError(t, err)
 	realInflater, ok := inflater.(*daisyInflater)
